feat: added conf page.
